07_difymcp_strandsagent_demo/strands-web-ui/requirements.txt
streamlit>=1.30.0
strands-agents>=0.1.1
mcp>=0.1.0
boto3>=1.26.0
pydub>=0.25.1
amazon-transcribe>=0.6.0
